Updated profile page and did minor corrections in paddings and widgets sizes on other pages
